api/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
anthropic==0.49.0
python-dotenv==1.0.1
pydantic==2.10.6
